+++ src/common/crc32.c
@@ -28,7 +28,7 @@ releases it into the public domain.
 
 #include <string.h>
 
-unsigned long crcTable[256] = {
+unsigned int crcTable[256] = {
 0x0, 0x77073096, 0xee0e612c, 0x990951ba, 0x76dc419, 0x706af48f, 0xe963a535, 0x9e6495a3,
 0xedb8832, 0x79dcb8a4, 0xe0d5e91e, 0x97d2d988, 0x9b64c2b, 0x7eb17cbd, 0xe7b82d07, 0x90bf1d91,
 0x1db71064, 0x6ab020f2, 0xf3b97148, 0x84be41de, 0x1adad47d, 0x6ddde4eb, 0xf4d4b551, 0x83d385c7,
@@ -69,14 +69,13 @@ unsigned long crcTable[256] = {
 //
 //
 ///////////////////////////////////////////////////////////////////////////////////////////////
-//unsigned long crc32(char *c, int len, unsigned long crcinit)
 
 //know length
 //ruteine der vi kjenner lengden
 unsigned int crc32boithonl(char *c, int len) {
 
-    unsigned long crcinit = 0;
-    register unsigned long crc;
+    unsigned int crcinit = 0;
+    register unsigned int crc;
     char     *e = c + len;
 
     crc = crcinit^0xFFFFFFFF;
@@ -96,12 +95,12 @@ unsigned int crc32boitho(char *c) {
 /*
 //You can test this module by uncoment this rutine, and compile.
 int main (int argc, char *argv[]) {
-	unsigned long crc;
+	unsigned int crc;
 
 	
 
 	crc = Bcrc32(argv[1]);
 
-	printf("%s crc %lu\n",argv[1],crc);
+	printf("%s crc %u\n",argv[1],crc);
 }
 */
